Update README.md
For some reason when I got home and changed the file path for the FUNCTIONS FILE to our local file path the script is broken.  I'm not sure where the issue is with it.  However, basic gist behind this script is that the worker chooses whether application received is new application, reapplication, or recertification.  Also requires worker to select how it is received and who submitted the application.  Then there are options for if there was expedited screening performed, if the client was given or mailed an appointment letter, and if the client refused a same day appointment.  Finally the script is set to see if the application is an initial application for SNAP, and if it is, then it will write the "received applications" note, and automatically pull up the expedited screening script and to have the worker complete that and will write that note as well.  This was set to only run if it is an initial application as expedited screening would not need to be performed if client is applying for recertification (currently active for SNAP).